+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * csirt_contact now may be derived from tld
     * web module now ignores invalid HTTPS certificates
     * form_names prints out more information about tags without `name` attribute
+    * base64 no more limited to UTF-8 
 * dialog
     * edit default e-mail templates from config submenu
     * delimiter now may contain tab character, input by "\t" or "tab"